requirements.txt

Flask
Pillow
flask
ipywidgets
jupyterlab
matplotlib
numpy
opencv-python
pandas
requests
scikit-image
scikit-learn
scipy
seaborn
tensorflow
tensorflow==2.15.0
torch
torchaudio  
torchvision
tqdm
werkzeugIPython
cv2
io
os
sklearn
warnings
